added regex to page search.js, get the right id link from the search field
